Unison.Merge

Documentation

data Diffblob libdep

Constructors

Diffblob

Fields

conflicts :: TwoWay (DefnsF (Map Name) TermReference TypeReference)

declNameLookups :: GThreeWay PartialDeclNameLookup DeclNameLookup

defns :: ThreeWay UnconflictedLocalDefnsView

defnsIds :: ThreeWay (DefnsF Set TermReferenceId TypeReferenceId)

diff :: DefnsF2 (Map Name) CombinedDiffOp Referent TypeReference

diffsFromLCA :: TwoWay (DefnsF3 (Map Name) DiffOp Synhashed Referent TypeReference)

hydratedNarrowedDefns :: Defns (Map TermReferenceId (Term Symbol Ann, Type Symbol Ann)) (Map TypeReferenceId (Decl Symbol Ann))

libdeps :: Updated (Map NameSegment libdep)

libdepsDiffs :: TwoWay (Map NameSegment (DiffOp libdep))

propagatedUpdates :: TwoWay (DefnsF (Map Name) (Updated Referent) (Updated TypeReference))

simpleRenames :: TwoWay (Defns SimpleRenames SimpleRenames)

unconflicts :: DefnsF Unconflicts Referent TypeReference

makeDiffblob :: (Eq libdep, Monad m) => DiffblobLog m -> (ThreeWay (DefnsF Set TermReferenceId TypeReferenceId) -> m (Defns (Map TermReferenceId (Term Symbol Ann, Type Symbol Ann)) (Map TypeReferenceId (Decl Symbol Ann)))) -> (ThreeWay (Set LabeledDependency) -> m (ThreeWay Names)) -> ThreeWay UnconflictedLocalDefnsView -> ThreeWay (Map NameSegment libdep) -> GThreeWay PartialDeclNameLookup DeclNameLookup -> m (Diffblob libdep)

data DiffblobLog (m :: Type -> Type)

Constructors

DiffblobLog

Fields

logDefns :: ThreeWay (DefnsF (Map Name) Referent TypeReference) -> m ()

logNarrowedDefns :: TwoWay (Updated (DefnsF (Map Name) Referent TypeReference)) -> m ()

logSynhashedNarrowedDefns :: TwoWay (GUpdated (DefnsF2 (Map Name) Synhashed Referent TypeReference) (DefnsF2 (Map Name) Synhashed Referent TypeReference)) -> m ()

logDiffsFromLCA :: TwoWay (DefnsF3 (Map Name) DiffOp Synhashed Referent TypeReference) -> m ()

logDiff :: DefnsF2 (Map Name) CombinedDiffOp Referent TypeReference -> m ()

emptyDiffblobLog :: forall (m :: Type -> Type). Applicative m => DiffblobLog m

data Mergeblob libdep

Constructors

Mergeblob

Fields

conflicts :: TwoWay (DefnsF (Map Name) TermReferenceId TypeReferenceId)

typecheckedFile :: Maybe (TypecheckedUnisonFile Symbol Ann)

unconflictedDefns :: DefnsF (Map Name) Referent TypeReference

uniqueTypeGuids :: TwoWay (Map Name Text)

unparsedFile :: Pretty ColorText

unparsedSoloFiles :: ThreeWay (Pretty ColorText)

data MergeblobError

Constructors

MergeblobError'ConflictedAlias (EitherWay (Defn (Name, Name) (Name, Name)))

MergeblobError'ConflictedBuiltin (Defn Name Name)

makeMergeblob :: Monad m => (ThreeWay (DefnsF Set TermReferenceId TypeReferenceId) -> m (Defns (Map TermReferenceId (Term Symbol Ann, Type Symbol Ann)) (Map TypeReferenceId (Decl Symbol Ann)))) -> (DefnsF Set TermReferenceId TypeReferenceId -> DefnsF Set TermReference TypeReference -> m (DefnsF Set TermReferenceId TypeReferenceId)) -> m (Updated Names) -> (DefnsF Set TermReference TypeReference -> m (TypeLookup Symbol Ann)) -> Diffblob libdep -> TwoWay Text -> m (Either MergeblobError (Mergeblob libdep))

Types

data CombinedDiffOp a

The combined result of two diffs on the same thing.

Constructors

CombinedDiffOp'Add !(EitherWayI a)

CombinedDiffOp'Delete !(EitherWayI a)

CombinedDiffOp'Update !(EitherWayI (Updated a))

CombinedDiffOp'Conflict !(TwoWay a)

data DiffOp a

A diff operation is one of:

An add (where nothing was)

A delete (of the thing that was)

An update (from old to new)

Constructors

DiffOp'Add !a

DiffOp'Delete !a

DiffOp'Update !(Updated a)

data EitherWay a

Alice exclusive-or Bob?

Constructors

Alice a

Bob a

data EitherWayI a

Alice inclusive-or Bob?

Constructors

OnlyAlice a

OnlyBob a

AliceAndBob a

data GThreeWay a b

Constructors

GThreeWay

Fields

lca :: a

alice :: b

bob :: b

data GUpdated a b

An updated thing.

Constructors

Updated

Fields

old :: a

new :: b

data LibdepDiffOp a

Constructors

AddLibdep !a

AddBothLibdeps !a !a

DeleteLibdep

data Rename

A "rename" is a venn partition of two non-empty sets of names: both set differences and the set intersection.

Invariant: the sets are all disjoint Invariant: it is not the case that adds and deletes are both empty

Constructors

Rename

Fields

adds :: Set Name

deletes :: Set Name

unchanged :: Set Name

data SimpleRenames

A "simple" rename is one that moves one name to another, where neither the before- nor after-name have any aliases.

Constructors

SimpleRenames

Fields

forwards :: !(Map Name Name)

backwards :: !(Map Name Name)

data Synhashed a

A small utility type that represents a syntactic-hashed thing.

The synhash itself is a lazy field so that we can avoid computing it in certain cases, just using laziness.

The Eq and Ord instances only compares syntactic hashes.

Constructors

Synhashed

Fields

hash :: Hash

value :: !a

data ThreeWay a

Constructors

ThreeWay

Fields

lca :: !a

alice :: !a

bob :: !a

data TwoOrThreeWay a

Constructors

TwoOrThreeWay

Fields

lca :: Maybe a

alice :: a

bob :: a

data TwoWay a

Constructors

TwoWay

Fields

alice :: a

bob :: a

data TwoWayI a

"Two-way inclusive".

Constructors

TwoWayI

Fields

alice :: a

bob :: a

both :: a

data Unconflicts v

Constructors

Unconflicts

Fields

adds :: !(TwoWayI (Map Name v))

deletes :: !(TwoWayI (Map Name v))

updates :: !(TwoWayI (Map Name v))

type Updated a = GUpdated a a