Unison.Merge.Narrow

Documentation

narrowDefns :: HasCallStack => GThreeWay PartialDeclNameLookup DeclNameLookup -> ThreeWay (DefnsF (Map Name) Referent TypeReference) -> TwoWay (Updated (DefnsF (Map Name) Referent TypeReference))

narrowDefnsTotal :: Updated DeclNameLookup -> Updated (DefnsF (Map Name) Referent TypeReference) -> Updated (DefnsF (Map Name) Referent TypeReference)

Like narrowDefns, but just between the LCA and a branch head, and for when the LCA is known not to have any type declarations with missing constructor names.