Unison.Merge.PreconditionViolation

Documentation

data PreconditionViolation

A reason that a merge could not be performed.

Constructors

ConflictedAliases !ProjectBranchName !Name !Name
ConflictedAliases branch foo bar: in project branch branch, foo and bar refer to different things, but at one time (in the LCA of another branch, in fact) they referred to the same thing.

ConflictedTermName !Name !(Set Referent)
ConflictedTermName name refs: name refers to 2+ referents refs.

ConflictedTypeName !Name !(Set TypeReference)
ConflictedTypeName name refs: name refers to 2+ type references refs.

ConflictInvolvingBuiltin !Name
ConflictInvolvingBuiltin name: name is involved in a conflict, but it refers to a builtin (on at least one side). Since we can't put a builtin in a scratch file, we bomb in these cases.

ConstructorAlias !(Maybe ProjectBranchName) !Name !Name
A second naming of a constructor was discovered underneath a decl's name, e.g.

Foo#Foo Foo.Bar0 Foo.Some.Other.Name.For.Bar0

If the project branch name is missing, it means the LCA is in violation.

DefnsInLib
There were some definitions at the top level of lib.*, which we don't like

MissingConstructorName !Name
This type name is missing a name for one of its constructors.

NestedDeclAlias !Name !Name
This type name is a nested alias, e.g. Foo.Bar.Baz which is an alias of Foo or Foo.Bar.

StrayConstructor !Name