Unison.Merge.Rename

Documentation

data Rename

A "rename" is a venn partition of two non-empty sets of names: both set differences and the set intersection.

Invariant: the sets are all disjoint Invariant: it is not the case that adds and deletes are both empty

Constructors

Rename

Fields

adds :: Set Name

deletes :: Set Name

unchanged :: Set Name

makeRenames :: Updated (Defns (BiMultimap (Synhashed Referent) Name) (BiMultimap (Synhashed TypeReference) Name)) -> DefnsF [] Rename Rename

data SimpleRenames

A "simple" rename is one that moves one name to another, where neither the before- nor after-name have any aliases.

Constructors

SimpleRenames

Fields

forwards :: !(Map Name Name)

backwards :: !(Map Name Name)

makeSimpleRenames :: DefnsF [] Rename Rename -> Defns SimpleRenames SimpleRenames