Unison.Merge

Documentation

data Mergeblob0 libdep

Constructors

Mergeblob0

Fields

defns :: ThreeWay (Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name))

libdeps :: ThreeWay (Map NameSegment libdep)

nametrees :: ThreeWay (Nametree (DefnsF (Map NameSegment) Referent TypeReference))

makeMergeblob0 :: ThreeWay (Nametree (DefnsF (Map NameSegment) Referent TypeReference)) -> ThreeWay (Map NameSegment libdep) -> Mergeblob0 libdep

data Mergeblob1 libdep

Constructors

Mergeblob1

Fields

conflicts :: TwoWay (DefnsF (Map Name) TermReference TypeReference)

declNameLookups :: TwoWay DeclNameLookup

defns :: ThreeWay (Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name))

diff :: DefnsF2 (Map Name) CombinedDiffOp Referent TypeReference

diffs :: TwoWay (DefnsF3 (Map Name) DiffOp Synhashed Referent TypeReference)

hydratedDefns :: ThreeWay (DefnsF (Map Name) (TermReferenceId, (Term Symbol Ann, Type Symbol Ann)) (TypeReferenceId, Decl Symbol Ann))

lcaDeclNameLookup :: PartialDeclNameLookup

libdeps :: Map NameSegment libdep

libdepsDiff :: Map NameSegment (LibdepDiffOp libdep)

unconflicts :: DefnsF Unconflicts Referent TypeReference

makeMergeblob1 :: forall libdep. Eq libdep => Mergeblob0 libdep -> ThreeWay (DefnsF (Map Name) (TermReferenceId, (Term Symbol Ann, Type Symbol Ann)) (TypeReferenceId, Decl Symbol Ann)) -> Either (EitherWay IncoherentDeclReason) (Mergeblob1 libdep)

data Mergeblob2 libdep

Constructors

Mergeblob2

Fields

conflicts :: TwoWay (DefnsF (Map Name) TermReferenceId TypeReferenceId)

coreDependencies :: TwoWay (DefnsF Set TermReference TypeReference)

declNameLookups :: TwoWay DeclNameLookup

defns :: ThreeWay (Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name))

hasConflicts :: Bool

hydratedDefns :: TwoWay (DefnsF (Map Name) (TermReferenceId, (Term Symbol Ann, Type Symbol Ann)) (TypeReferenceId, Decl Symbol Ann))

libdeps :: Map NameSegment libdep

soloUpdatesAndDeletes :: TwoWay (DefnsF Set Name Name)

unconflicts :: DefnsF Unconflicts Referent TypeReference

data Mergeblob2Error

Constructors

Mergeblob2Error'ConflictedAlias (EitherWay (Defn (Name, Name) (Name, Name)))

Mergeblob2Error'ConflictedBuiltin (Defn Name Name)

makeMergeblob2 :: Mergeblob1 libdep -> Either Mergeblob2Error (Mergeblob2 libdep)

data Mergeblob3

Constructors

Mergeblob3

Fields

libdeps :: Names

stageOne :: DefnsF (Map Name) Referent TypeReference

stageTwo :: DefnsF (Map Name) Referent TypeReference

uniqueTypeGuids :: Map Name Text

unparsedFile :: Pretty ColorText

makeMergeblob3 :: Mergeblob2 libdep -> TwoWay (DefnsF Set TermReferenceId TypeReferenceId) -> Names -> TwoWay Text -> Mergeblob3

data Mergeblob4

Constructors

Mergeblob4

Fields

dependencies :: DefnsF Set TermReference TypeReference

file :: UnisonFile Symbol Ann

makeMergeblob4 :: Mergeblob3 -> Either (Err Symbol) Mergeblob4

data Mergeblob5

Constructors

Mergeblob5

Fields

file :: TypecheckedUnisonFile Symbol Ann

makeMergeblob5 :: Mergeblob4 -> TypeLookup Symbol Ann -> Either (Seq (Note Symbol Ann)) Mergeblob5

Decl coherency checks

data PartialDeclNameLookup

Like a DeclNameLookup, but "partial" / more lenient - because we don't require the LCA of a merge to have a full DeclNameLookup.

Constructors

PartialDeclNameLookup

Fields

constructorToDecl :: !(Map Name Name)

declToConstructors :: !(Map Name [Maybe Name])

data IncoherentDeclReason

Constructors

IncoherentDeclReason'ConstructorAlias !Name !Name !Name
A second naming of a constructor was discovered underneath a decl's name, e.g.

Foo#Foo Foo.Bar0 Foo.Some.Other.Name.For.Bar0

IncoherentDeclReason'MissingConstructorName !Name

IncoherentDeclReason'NestedDeclAlias !Name !Name
A second naming of a decl was discovered underneath its name, e.g.

Foo#Foo Foo.Bar#Foo

IncoherentDeclReason'StrayConstructor !TypeReferenceId !Name

checkDeclCoherency :: HasCallStack => Nametree (DefnsF (Map NameSegment) Referent TypeReference) -> Map TypeReferenceId Int -> Either IncoherentDeclReason DeclNameLookup

lenientCheckDeclCoherency :: Nametree (DefnsF (Map NameSegment) Referent TypeReference) -> Map TypeReferenceId Int -> PartialDeclNameLookup

A lenient variant of checkDeclCoherency - so lenient it can't even fail! It returns partial decl name lookup, which doesn't require a name for every constructor, and allows a constructor with a nameless decl.

This function exists merely to extract a best-effort name mapping for the LCA of a merge. We require Alice and Bob to have coherent decls, but their LCA is out of the user's control and may have incoherent decls, and whether or not it does, we still need to compute *some* syntactic hash for its decls.

data IncoherentDeclReasons

Constructors

IncoherentDeclReasons

Fields

constructorAliases :: ![(Name, Name, Name)]

missingConstructorNames :: ![Name]

nestedDeclAliases :: ![(Name, Name)]

strayConstructors :: ![(TypeReferenceId, Name)]

checkAllDeclCoherency :: forall m. Monad m => (TypeReferenceId -> m Int) -> Nametree (DefnsF (Map NameSegment) Referent TypeReference) -> m (Either IncoherentDeclReasons DeclNameLookup)

Like checkDeclCoherency, but returns info about all of the incoherent decls found, not just the first.

Types

data CombinedDiffOp a

The combined result of two diffs on the same thing.

Constructors

CombinedDiffOp'Add !(EitherWayI a)

CombinedDiffOp'Delete !(EitherWayI a)

CombinedDiffOp'Update !(EitherWayI (Updated a))

CombinedDiffOp'Conflict !(TwoWay a)

data DiffOp a

A diff operation is one of:

An add (where nothing was)

A delete (of the thing that was)

An update (from old to new)

Constructors

DiffOp'Add !a

DiffOp'Delete !a

DiffOp'Update !(Updated a)

data EitherWay a

Alice exclusive-or Bob?

Constructors

Alice a

Bob a

data EitherWayI a

Alice inclusive-or Bob?

Constructors

OnlyAlice a

OnlyBob a

AliceAndBob a

data LibdepDiffOp a

Constructors

AddLibdep !a

AddBothLibdeps !a !a

DeleteLibdep

data Synhashed a

A small utility type that represents a syntactic-hashed thing.

The Eq and Ord instances only compares syntactic hashes.

Constructors

Synhashed

Fields

hash :: !Hash

value :: !a

data ThreeWay a

Constructors

ThreeWay

Fields

lca :: !a

alice :: !a

bob :: !a

data TwoOrThreeWay a

Constructors

TwoOrThreeWay

Fields

lca :: Maybe a

alice :: a

bob :: a

data TwoWay a

Constructors

TwoWay

Fields

alice :: a

bob :: a

data TwoWayI a

"Two-way inclusive".

Constructors

TwoWayI

Fields

alice :: a

bob :: a

both :: a

data Unconflicts v

Constructors

Unconflicts

Fields

adds :: !(TwoWayI (Map Name v))

deletes :: !(TwoWayI (Map Name v))

updates :: !(TwoWayI (Map Name v))

data Updated a

An updated thing.

Constructors

Updated

Fields

old :: a

new :: a