Unison.Merge.Internal.Types

Description

Internal types module to house types that would require mutual recursion at the module level if defined separately

Documentation

data ThreeWay a

Constructors

ThreeWay

Fields

lca :: !a

alice :: !a

bob :: !a

data TwoOrThreeWay a

Constructors

TwoOrThreeWay

Fields

lca :: Maybe a

alice :: a

bob :: a